+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 [Full changelog](https://github.com/mozilla/glean/compare/v22.0.0...master)
 
+* Python bindings:
+
+  * Support for events was added.
+
 # v22.0.0 (2019-12-05)
 
 [Full changelog](https://github.com/mozilla/glean/compare/v21.3.0...22.0.0)